+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Do not generate proxy env secret for non private clusters.
+
 ## [2.20.0] - 2024-02-12
 
 ### Changed